Allow users to create & destroy their microposts
Actions like new or edit in the Microposts controller won't be needed
because the interface to the Microposts resource will run mainly
through the Profile and Home pages.
* Adjust routes.rb and limit resources for microposts
* Adjust test/controllers/microposts_controller_test.rb and add tests
to ensure that users are logged in for the create & destroy actions.
* Adjust controllers/application_controller.rb, move the logged_in_user
method from the Users controller into the Application controller, since
it will be used by more than a single controller.
* Adjust controllers/microposts_controller.rb
* Adjust views/static_pages/home_page.html.erb
* Create a partial views/shared/_user_info.html.erb
* Create a partial views/shared/_micropost_form.html.erb
* Adjust controllers/static_pages_controller.rb & define @micropost.
* Redefine error_messages partial to ensure that the micropost_form
partial works correctly. There will be a knock-on effect that will be
revealed by running your test suite. Adjust views/users/_form.html.erb,
as well as views/password_resets/edit.html.erb